SDK changes and advancements!
I removed the web implementation of the SDK as this will be implemented
in the right repository. Currently this is still in development but it
is planed that the interfaces and abstract classes of sdk will be moved
into their own directories and then imported (e.g. via git submodules).
The implementation will then live inisde of the respective repo.
I also added the extensions resolver it still needs the validators
though