Reworked the extension host controller
Mainly reworked the extension host controller. Moved the IFramController
form inside of the ExtensionHostController into the ExtensionService as
it was completly useless to put it there. Hooked everything up with the
identifiers as their refernces and thus simplified the spaces logic. I
also added the EventTargetController to the ExtensionHost. Big commit!
Never work when you are sleepy!